Conflicts:
	src/CCStats.cs
	src/actions/action_grid/CCGrid3DAction.cs
	src/menu_nodes/CCMenuItemLabel.cs

+++ src/label_nodes/CCLabel-Android.cs
@@ -23,7 +23,7 @@ namespace CocosSharp
             var contentScaleFactorWidth = CCLabel.DefaultTexelToContentSizeRatios.Width;
             var contentScaleFactorHeight = CCLabel.DefaultTexelToContentSizeRatios.Height;
 
-            textDef.FontSize *= (int)contentScaleFactorWidth;
+            textDef.FontSize *= contentScaleFactorWidth;
             textDef.Dimensions.Width *= contentScaleFactorWidth;
             textDef.Dimensions.Height *= contentScaleFactorHeight;
